Teste carregando usuario

+++ backend/src/main/java/com/mariana/dscommerce1/services/ProductService.java
@@ -4,7 +4,6 @@ import javax.persistence.EntityNotFoundException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -58,18 +57,18 @@ public class ProductService {
         }
     }
     @Transactional(propagation = Propagation.SUPPORTS)
-    public void delete(Long id){
-        try{
-            repository.deleteById(id);
-        }
-        catch (EmptyResultDataAccessException e){
-            throw new ResourcesNotFoundException("Recurso não encontrado");
-        }
-        catch (DataIntegrityViolationException e){
-            throw new DatabaseException("Falha de integridade referencial");
-
-        }
+    public void delete(Long id) {
+    	if (!repository.existsById(id)) {
+    		throw new ResourcesNotFoundException("Recurso não encontrado");
+    	}
+    	try {
+    		repository.deleteById(id);
+    	}
+    	catch (DataIntegrityViolationException e) {
+    		throw new DatabaseException("Falha de integridade referencial");
+    	}
     }
+   
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
         entity.setName(dto.getName());
